backend/requirements.txt
fastapi==0.132.0
joblib==1.4.2
pandas==3.0.1
pydantic==2.12.5
wandb==0.25.0
numpy==2.4.2
pybaseball==2.2.7
scikit-learn==1.6.1
scipy==1.17.1
xgboost==3.2.0
boto3
pydantic
datetime
botocore
uuid
MLB-StatsAPI==1.9.0